Made sure GPUreadStatus() is being cast to an unsigned long

+++ stationary_test_main.c
@@ -55,9 +55,9 @@ int main(int argc, char* argv[]) {
 	GPUdisplayText("You are now running Stationary");
 	updateDisplay();
 	printf("You are now running stationary on display %lu\n", disp);
-    printf("The GPU has the following status: 0x%08lx\n", GPUreadStatus());
+    printf("The GPU has the following status: 0x%08lx\n", (unsigned long)GPUreadStatus());
     S_do_gpu_startup();
-    printf("The GPU has the following status: 0x%08lx\n", GPUreadStatus());
+    printf("The GPU has the following status: 0x%08lx\n", (unsigned long)GPUreadStatus());
     S_draw_tri(0, 200, 100, 0, 200, 200, 0x7F, 0x00, 0x00);
     updateDisplay();
     while(1);
